fix : onCreate 에서 navHost 파일 분리
